Deleting all categories related

+++ src/services/categories/DeleteCategoryService.ts
@@ -1,4 +1,5 @@
 import { Category } from '@database/models/Category';
+import AppError from '@errors/AppError';
 import { getRepository } from 'typeorm';
 
 interface Request {
@@ -9,11 +10,29 @@ class DeleteCategoryService {
 	public async execute({ id }: Request): Promise<void> {
 		const categoriesRepository = getRepository(Category);
 
-		await categoriesRepository.delete(id);
+		const category = await categoriesRepository.findOne(id);
 
-		await categoriesRepository.delete({
+		if (!category) {
+			throw new AppError('Essa categoria não existe.');
+		}
+
+		await this.deleteChildren(id);
+
+		await categoriesRepository.delete(id);
+	}
+
+	private async deleteChildren(id: string) {
+		const categoriesRepository = getRepository(Category);
+
+		const category = await categoriesRepository.findOne({
 			parentId: id,
 		});
+
+		if (category) {
+			await this.deleteChildren(category.id);
+		}
+
+		await categoriesRepository.delete(id);
 	}
 }
